Add JUCE threading patches that fix juce::Thread issue wrt ClassLoader.
The patch has been around for a while as
https://atsushieno.github.io/2022/03/16/juce-android-threads.html

Since JUCE never fixes their threading issue and therefore it is basically
useless, it should be locally applied.

There are actually two patches; one is for JUCE6 and the other is for JUCE7.

One will have to be automatically applied (TODO).