AudioPluginFormat: rewrite instancing to fix threading issue.
It is to fix https://github.com/atsushieno/aap-juce/issues/35

First, createInstanceAsync() is deprecated in AAP.

The instancing involves `Context.bindService()` which should not be blocking
on the main thread. Since `AudioPluginHostHelper.ensureBinderConnected()`
is a synchronous function and it blocks until `bindService()` result is
returned from the Service, it must not run on the main thread.
Thus we launch another native thread (`juce::Thread`) to not block the
main thread now.