Add JUCE patch that disables DetachCurrentThread().
The patch is actually quite hacky as explained in the patch - the problem
is that the thread that is originated from java.lang.Thread should not
involve DetachCurrentThread().
The patch simply disables it. It looks leaky, but the only thread that
misses DetachCurrentThread() is, to my understanding, the main thread.