migrate to the new-plugin-buffer-api of the latest aap-core.
Though I figured that there is a major problem within our plugin protocol (AIDL) -
we need to pass the actual frame count on every `process()` call.
JUCE AudioPluginHost passes inconsistent frame count every time, and that actually
matches what Oboe does. In or MidiDeviceService implementation we enqueue the
audio inputs and flush only when it is filled. We could do the same in aap-juce,
but now it would make more sense to try to let binder process the audio buffer
in flexible size.