Do not allow plugin's standalone filter app to automatically open MIDI devices.
JUCE on Android is too immature so that we never allow aap-juce apps to
automatically open MIDI devices.

JUCE does not even understand which devices are input and which are output.
So, even if they think it is kindful to automatically open ALL MIDI INPUTS,
it actually opens ALL MIDI OUTPUTS too.

What is the implication? It automatically opens all the MidiDeviceServices
that we offer! It is significantly blocking any other audio apps.

What is even worse is, no matter what Activity we specify for Main and
Launcher on AndroidManifest.xml, JUCE initializer forcibly launches
JuceActivity. Therefore it ALWAYS opens the MIDI devices.

However, to the last resort, we can disable this behavior by specifying
JUCE_DONT_AUTO_OPEN_MIDI_DEVICES_ON_MOBILE=1.
We make it mandatory to disable it at all.
The default "auto-open MIDI device" feature should go to hell.

No mercy.