BREAKING CHANGE: rename directory `modules` to `aap-modules`
This impacts on ALL existing Projucer-based apps: their `override.*.jucer`
references `modules` directory in each juceaap module paths. They have to be
updated to point to `aap-modules`.

The reason why we make this change is aap-juce-obxd: it contains a directory
"Modules" within the source tree, and it conflicts with our symlink... on
case-insensitive filesystems.

While the actual files in that directory are not used by ourselves anyways,
we should make consistent build system across multiple projects as long as
possible. "modules" is certainly a bad name that any JUCE plugins could
potentially have, not just OB-Xd.